feat: avatar component and delete avatar
Extracted avatar upload and delete logic into a new Avatar component for the settings page. Updated user context and service to use a new deleteAvatarByCookie function with proper API endpoint and method. Adjusted localStorage usage for avatar image URL and improved error handling. Also added a placeholder for password change functionality in settings.
